eatgo/eatgo-api/src/main/java/kr/co/fascampus/eatgo/application/ReviewService.java
package kr.co.fascampus.eatgo.application;

import kr.co.fascampus.eatgo.domain.Review;
import kr.co.fascampus.eatgo.domain.ReviewRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class ReviewService {

    @Autowired
    private final ReviewRepository reviewRepository;

    public ReviewService(ReviewRepository reviewRepository) {
        this.reviewRepository = reviewRepository;
    }

    public Review addReview(Review review) {
        return reviewRepository.save(review);
    }
}
